@@ -1,7 +1,7 @@
 # From title_pattern.py
 
 
-import sys
+from collections import Counter, defaultdict
 from typing import Any
 
 from oldnyc.geocode import grid
@@ -15,8 +15,18 @@ from oldnyc.geocode.geocode_types import (
 )
 from oldnyc.item import Item
 
+# (coder, event) -> count
+counts = defaultdict[str, Counter[str]](Counter)
 
-def locate_with_osm(r: Item, loc: Locatable) -> Point | None:
+
+def round_pt(pt: Point) -> Point:
+    lat, lng = pt
+    return round(lat, 7), round(lng, 7)
+
+
+def locate_with_osm(
+    r: Item, loc: Locatable, coder: str, grid_geocoder: grid.GridGeocoder
+) -> Point | None:
     """Extract a location from a Locatable, without going to Google."""
     if isinstance(loc, LatLngLocation):
         return round(loc.lat, 7), round(loc.lng, 7)
@@ -25,21 +35,18 @@ def locate_with_osm(r: Item, loc: Locatable) -> Point | None:
     # Must be an intersection
     assert isinstance(loc, IntersectionLocation)
 
-    str1 = loc.str1
-    str2 = loc.str2
     boro = loc.boro
-    if boro not in ("Manhattan", "New York"):
-        return None
+    boro = "Manhattan" if boro == "New York" else boro
     try:
-        ave, street = grid.parse_street_ave(str1, str2)
-        pt = grid.code(ave, street)
+        counts[coder]["grid: attempt"] += 1
+        pt = grid_geocoder.geocode_intersection(loc.str1, loc.str2, boro, r.id)
     except ValueError:
         # sys.stderr.write(f"grid fail\t{r.id}\t{loc}\n")
         return None
     if not pt:
         return None
-    lat, lng = pt
-    return round(float(lat), 7), round(float(lng), 7)  # they're numpy floats
+    counts[coder]["grid: success"] += 1
+    return round_pt(pt)
 
 
 def get_address_for_google(loc: Locatable) -> str:
@@ -52,7 +59,7 @@ def get_address_for_google(loc: Locatable) -> str:
 
 
 def extract_point_from_google_geocode(
-    geocode: dict[str, Any], loc: Locatable, record: Item
+    geocode: dict[str, Any], loc: Locatable, record: Item, coder: str
 ) -> Point | None:
     if isinstance(loc, LatLngLocation):
         return loc.lat, loc.lng
@@ -64,18 +71,21 @@ def extract_point_from_google_geocode(
         raise ValueError()
 
     if not pt:
+        counts[coder]["google: fail"] += 1
         return None
     lat, lng = pt
     geocode_boro = point_to_borough(lat, lng)
     boro = loc.boro if loc.boro != "New York" else "Manhattan"
     if geocode_boro != boro:
         # self.n_boro_mismatch += 1
-        sys.stderr.write(
-            f"Borough mismatch: {record.id}: {loc.source} geocoded to {geocode_boro} not {boro}\n"
-        )
+        # sys.stderr.write(
+        #     f"Borough mismatch: {record.id}: {loc.source} geocoded to {geocode_boro} not {boro}\n"
+        # )
+        counts[coder]["google: boro mismatch"] += 1
         return None
-    # self.n_success += 1
-    return round(float(lat), 7), round(float(lng), 7)
+    # TODO: track hits by locatable type
+    counts[coder]["google: success"] += 1
+    return round_pt(pt)
 
 
 def get_lat_lng_from_geocode(geocode: dict[str, Any], desired_types: list[str]) -> Point | None:
